0.5.0 - create game enemies (witch). Bug fix

+++ app/src/main/java/com/example/corgirun/objects/animation/AnimationFire.java
@@ -45,7 +45,7 @@ public class AnimationFire {
 		}
 	}
 
-	public void drawingAnimation(GraphicsPuz graphicsPuz, int x, double y) {
+	public void drawingAnimation(GraphicsPuz graphicsPuz, double x, double y) {
 		graphicsPuz.drawTexture(sprite, x, y);
 	}
